Refactor: extract shared plugin helpers, split oversized plugins
Behavior-preserving refactor across the plugin tree:

- Add plugin::oauth::ClientCredentials, a shared OAuth2 client-credentials
  token cache used by the spotify and twitch plugins, replacing their
  duplicated Token/RwLock/get_token implementations. AuthStyle covers the
  Basic-header (Spotify) vs form-body (Twitch) difference. Spotify's token
  refresh now uses reqwest's built-in basic_auth, dropping the base64
  dependency, and calls error_for_status() so failed auth surfaces as an
  HTTP error instead of a JSON decode error.

- Add utils::parse_json, deduplicating the serde_path_to_error decode +
  log-body-on-failure idiom repeated in the pornhub, tvmaze, reddit, and
  youtube plugins. Also remove spotify's unused Error::Json variant.

- Add plugin::prefixed for the standard colored "> Name:" IRC message
  prefix, adopted only by the plugins whose format strings were
  byte-identical (github, rust_playground, spotify, twitch); all other
  prefix variants are left untouched.

- Split the three largest plugin files into module directories following
  the existing kagi/ layout:
  reddit.rs (688) -> reddit.rs + reddit/{api,types}.rs
  youtube.rs (575) -> youtube.rs + youtube/{api,types,urls}.rs
  isitopen.rs (522) -> isitopen.rs + isitopen/{hours,types}.rs

No changes to logic, IRC output bytes, log messages, or error variants
(other than the two spotify items noted above). All existing tests pass
in their new locations.
